pyproject.toml
[project]
name = "augview"
version = "0.1.0"
description = "Image Augmentation Pipeline Visualizer"
authors = [
    {name = "AugView Team"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["image", "augmentation", "visualization", "deep-learning", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "websockets>=11.0",
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
albumentations = ["albumentations>=1.3.0"]
torchvision = ["torchvision>=0.15.0"]
all = [
    "albumentations>=1.3.0",
    "torchvision>=0.15.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.scripts]
augview = "augview.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["augview*"]

[tool.setuptools.package-data]
augview = ["frontend/dist/**/*"]
